+++ packages/rum_sdk/lib/rum_flutter.dart
@@ -12,7 +12,7 @@ import 'package:rum_sdk/src/models/span_record.dart';
 import 'package:rum_sdk/src/tracing/tracer_provider.dart';
 import 'package:rum_sdk/src/transport/batch_transport.dart';
 import 'package:rum_sdk/src/util/generate_session.dart';
-import 'package:rum_sdk/src/util/session_attributes_provider.dart';
+import 'package:rum_sdk/src/device_info/session_attributes_provider.dart';
 
 Timer? timer;
