+++ packages/rum_sdk/lib/src/device_info/platform_info_provider.dart
