+++ packages/rum_sdk/test/src/device_info/device_id_provider_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:rum_sdk/src/util/device_id_provider.dart';
+import 'package:rum_sdk/src/device_info/device_id_provider.dart';
 import 'package:rum_sdk/src/util/uuid_provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
